feat: add the methods to the builder trait, update code that was broken.
